Added code in notebook to concatenate and unstack values.
The notebook now produces a DataFrame with all criteria values, with one column
per criterion. We still need to add the full list of criteria, and to do
prepprocessing (aggregate sublevels to missing aggregate variables if desired,
such as the "Emissions|CO2" that is missing in some models, and add/scale with
GWPs and sum to get total GHG emissions if needed).